kvserver: add change executor
The store rebalancer, calls applies it's own changes, blocking on each.
In order to simulate the store rebalancer, it is necessary to have an
implementation of these changer functions. This patch introduces a
change executor interface which provides an abstracton that the real
code and the simulator may implement.

Release note: None