storage: don't return WriteTooOld on inline DeleteRange that hits non-inline value
Fixes #101440.

This commit resolves a condition where an inline DeleteRange request
could return a WriteTooOld error when encountering a non-inline value,
instead of returning the expected assertion error. This could trigger a
fatal error in `tryBumpBatchTimestamp`. See the corresponding issue for
more details.

Note that this is not resolving any condition that could cause an inline
DeleteRange request to hit a non-inline value, it's just ensuring that
if corruption creates such a condition, we don't fatal.

Release note: None