kvserver: validate returned lb split key
It is possible for the split key returned from the load based splitter
to be the start key of the range. This can occur when one row is
particularly hot and the row uses column families. In such cases,
although the load based finder would track 20 unique keys, with
different suffixes, the resulting split key returned would always be the
same prefix.

e.g. for YCSB with column families:

We wish to split at:

/Table/104/2/"user2933389304617401955"/7/1

However the returned, safe SQL key is:

/Table/104/2/‹"user2933389304617401955"›

Which is the only possible key returned for the range.

This commit doesn't address the same key being returned, instead it
validates that the key isn't equal to the start key before the split
queue begins processing the split.

Resolves: #102136

Release note: None