allocator: log when a lease candidate is invalid
It can be puzzling to figure out why a replica wasn't a potential
candidate for a lease transfer, when expected to be. This concern is
mostly relevant to debugging tests or investigations.

Log (v=5) a message when a candidate replica is not a valid transfer
target.

Epic: none

Resolves: #105066

Release note: None