kvserver: deflake lease preferences during outage
Previously, it was possible for a soon to be dead replica, to acquire
the range lease in the `TestLeasePreferencesDuringOutage` test. The
acquired lease would be expiration based, disallowing the intended
leaseholder from acquiring the lease.

This patch disables expiration based lease transfers, deflaking the
test.

Resolves: #105101
Epic: none
Release note: None